+++ build.test/FeatureExtenssions.cs
@@ -1,0 +1,43 @@
+using System.Text.Json;
+using Nuke.Common.IO;
+
+namespace build.test;
+
+public static class FeatureExtenssions
+{
+	public static Tag GetTag(this Feature feature) => new($"feature_{feature}");
+
+	public static AbsolutePath GetRoot(
+		this Feature feature,
+		AbsolutePath projectRoot) => projectRoot
+			/ "features"
+			/ "src"
+			/ feature;
+
+	public static AbsolutePath GetConfig(
+		this Feature featureName,
+		AbsolutePath projectRoot)
+		=> featureName.GetRoot(projectRoot)
+			/ "devcontainer-feature.json";
+
+	public static string GetRelativePathToConfig(this Feature feature)
+		=> Path.Combine("features", "src", feature, "devcontainer-feature.json");
+
+	public static async Task<string?> GetVersion(
+		this Feature feature,
+		AbsolutePath projectRoot)
+	{
+		var featureConfig = feature.GetConfig(projectRoot);
+		using var fileStream = File.OpenRead(featureConfig);
+		var document = await JsonDocument.ParseAsync(fileStream);
+
+		return document.RootElement.GetProperty("version").GetString();
+	}
+
+	public static void CreateTempFile(
+		this Feature feature,
+		AbsolutePath root)
+	{
+		using var _ = File.Create(feature.GetRoot(root) / $"tmp_{Guid.NewGuid():N}");
+	}
+}
