+++ build.test/Tag.cs
@@ -1,0 +1,11 @@
+namespace build.test;
+
+public sealed record Tag(string Name)
+{
+	public override string ToString()
+	{
+		return Name;
+	}
+
+	public static implicit operator string(Tag tag) => tag.ToString();
+}
