+++ build.test/VersionExtenssions.cs
@@ -1,0 +1,17 @@
+namespace build.test;
+
+public static class VersionExtenssions
+{
+	public static Version IncrementMajor(this Version version) => version with
+	{
+		Major = version.Major + 1,
+		Minor = 0,
+		Build = 0
+	};
+
+	public static Version IncrementMinor(this Version version) => version with
+	{
+		Minor = version.Minor + 1,
+		Build = 0
+	};
+}
